drlead/pyproject.toml
[project]
name = "drlead"
version = "0.1.0"
description = "lead generator"
authors = [
    {name = "Priyanshu Jha",email = "priyanshu.k4444@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pymongo (>=4.11.2,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "flask-pymongo (>=3.0.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
